Tools: Further fix for bug http://bugs.activestate.com/show_bug.cgi?id=88571
(integrated from komodo (trunk) change 58202 by ericp)

git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@6566 b87d16e7-d29f-4ee1-b5ec-817d645d335f

+++ src/toolbox/koToolbox2Components.py
@@ -557,7 +557,7 @@ class KoToolbox2Service(object):
             self.unregisterUserToolbox(project.url)
             self.notifyToolboxTopLevelViewChanged()
         toolsDir = join(projectDir, koToolbox2.PROJECT_TARGET_DIRECTORY)
-        if not os.listdir(toolsDir):
+        if exists(toolsDir) and not os.listdir(toolsDir):
             try:
                 os.rmdir(toolsDir)
             except:
